requirements.txt
streamlit>=1.30.0
supabase>=2.0.0
Pillow>=10.0.0
pytesseract>=0.3.10
groq>=0.4.0
